refactor(kotest): separate build-time emitter from test-time runtime
Two architecture fixes on the kotest integration module, driven by a
deep-modules / clear-interfaces review. Behavior-preserving; jvmTest green.

- Kill the Response<NNN> naming-convention leak: EndpointReflection,
  EndpointCallBuilder and EndpointShape each encoded "variant class =
  Response + 3-digit status" independently. Centralize into one internal
  ResponseVariantNaming (className/statusOf); all three defer to it.

- Split the overloaded extension/ package by lifecycle. It fused the
  build-time IR emitter (deps: compiler + ir) with the test-time Kotest
  framework hooks (deps: kotest.core) under one word ("extension" meaning
  both IrExtension and TestCaseExtension). Move the 10 internal codegen
  helpers (*DslFile, *Shape, RecordBuilder, KotlinIdentifier,
  KotlinTypeMapper, ValueSetter) into a new emit/ package; the runtime
  Kotest extensions and the public KotestDslExtension stay in extension/,
  so no public FQN changes. KotestDslExtension now imports the *DslFile
  builders from emit/.
