+++ src/integration/kotest/src/commonMain/kotlin/community/flock/wirespec/integration/kotest/emit/ValueSetter.kt
@@ -1,4 +1,4 @@
-package community.flock.wirespec.integration.kotest.extension
+package community.flock.wirespec.integration.kotest.emit
 
 import community.flock.wirespec.ir.core.StructBuilder
 import community.flock.wirespec.ir.core.Visibility
